% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_engine_cpp.R
\name{ply_fen_parse}
\alias{ply_fen_parse}
\title{Parse a FEN string into a ChessState list}
\usage{
ply_fen_parse(fen)
}
\arguments{
\item{fen}{A FEN string.}
}
\value{
A named list of class \code{ChessState} representing the board state.
}
\description{
Parse a FEN string into a ChessState list
}
\examples{
state <- ply_fen_parse("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1")
state$sideToMove  # 0 = white
}
