% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_engine_cpp.R
\name{ply_fen_serialize}
\alias{ply_fen_serialize}
\title{Convert a ChessState list back to a FEN string}
\usage{
ply_fen_serialize(state)
}
\arguments{
\item{state}{A \code{ChessState} list from \code{ply_fen_parse}.}
}
\value{
A FEN string.
}
\description{
Convert a ChessState list back to a FEN string
}
\examples{
state <- ply_game_init()
ply_fen_serialize(state)
}
