% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chess_engine_cpp.R
\name{ply_is_checkmate}
\alias{ply_is_checkmate}
\title{Test whether the position is checkmate}
\usage{
ply_is_checkmate(state)
}
\arguments{
\item{state}{A \code{ChessState} list.}
}
\value{
\code{TRUE} if the side to move is checkmated.
}
\description{
Test whether the position is checkmate
}
\examples{
# Scholar's Mate — black is checkmated
s <- ply_fen_parse(
  "r1bqkb1r/pppp1Qpp/2n2n2/4p3/2B1P3/8/PPPP1PPP/RNB1K1NR b KQkq - 0 4"
)
ply_is_checkmate(s)  # TRUE
}
