% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ply-package.R
\docType{package}
\name{ply-package}
\alias{ply}
\alias{ply-package}
\title{ply: A Bitboard Chess Engine for R}
\description{
Provides a fully legal chess move generator and game engine
implemented in C++17 via Rcpp. The package exposes two layers:
}
\details{
\itemize{
  \item A high-level R API (\code{ply_*} functions) for everyday use.
  \item An internal Rcpp bridge (\code{cpp_*} functions) used by the exported
  R helpers and test infrastructure.
}
}
\author{
\strong{Maintainer}: Qusai Jouda \email{qjouda@gmu.edu}

}
\keyword{internal}
